+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## swamp v0.12.0
 
 * `-alias-config`: generate lowercase profile names and aliases
+* `-export-profile` and `-export-file`: removed. Please use `-exec` instead. E.g. `-exec bash`.
 
 ## swamp v0.11.0